Adjustments based on PR feedback including factor out of types, typing changes removal of unnecessary imports
